+++ lib/ActionType.php
@@ -1,0 +1,9 @@
+<?php
+
+enum ActionType: string
+{
+    case TOGGLESTATUS = "togglestatus";
+    case ADD = "add";
+    case EDIT = "edit";
+    case DELETE = "delete";
+}