+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+
+## Version 1.1.0 -> 24.02.2022
+added table management
+
+
+## Version 1.0.5 -> 15.12.2021
+added database connection
+
+
 ## Version 1.0.0 -> 26.11.2021
 
 init